# api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Endpoints for LocalEats",
        "title": "LocalEats API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/kitchen": {
            "post": {
                "description": "Create a new kitchen with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kitchen"
                ],
                "summary": "Create a new kitchen",
                "parameters": [
                    {
                        "description": "Kitchen creation details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gateway_genprotos_auth_pb.CreateKitchenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gateway_genprotos_auth_pb.CreateKitchenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/kitchen/list": {
            "post": {
                "description": "Get a list of kitchens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kitchen"
                ],
                "summary": "List kitchens",
                "parameters": [
                    {
                        "description": "List kitchens request parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gateway_genprotos_auth_pb.ListKitchensRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gateway_genprotos_auth_pb.ListKitchensResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/kitchen/search": {
            "post": {
                "description": "Search for kitchens based on provided criteria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kitchen"
                ],
                "summary": "Search kitchens",
                "parameters": [
                    {
                        "description": "Search criteria",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gateway_genprotos_auth_pb.SearchKitchensRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gateway_genprotos_auth_pb.SearchKitchensResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/kitchen/{kitchen_id}": {
            "get": {
                "description": "Get details of a specific kitchen by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kitchen"
                ],
                "summary": "Get kitchen details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kitchen ID",
                        "name": "kitchen_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gateway_genprotos_auth_pb.GetKitchenResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing kitchen's details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "kitchen"
                ],
                "summary": "Update a kitchen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Kitchen ID",
                        "name": "kitchen_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated kitchen details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/gateway_genprotos_auth_pb.UpdateKitchenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gateway_genprotos_auth_pb.UpdateKitchenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gateway_genprotos_auth_pb.CreateKitchenRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cuisine_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "gateway_genprotos_auth_pb.CreateKitchenResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "cuisine_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "kitchen_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                }
            }
        },
        "gateway_genprotos_auth_pb.GetKitchenResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "cuisine_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "kitchen_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "total_orders": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "gateway_genprotos_auth_pb.KitchenListItem": {
            "type": "object",
            "properties": {
                "cuisine_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "total_orders": {
                    "type": "integer"
                }
            }
        },
        "gateway_genprotos_auth_pb.ListKitchensRequest": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "gateway_genprotos_auth_pb.ListKitchensResponse": {
            "type": "object",
            "properties": {
                "kitchens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gateway_genprotos_auth_pb.KitchenListItem"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "gateway_genprotos_auth_pb.SearchKitchensRequest": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "gateway_genprotos_auth_pb.SearchKitchensResponse": {
            "type": "object",
            "properties": {
                "kitchens": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gateway_genprotos_auth_pb.KitchenListItem"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "gateway_genprotos_auth_pb.UpdateKitchenRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "kitchen_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "gateway_genprotos_auth_pb.UpdateKitchenResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "cuisine_type": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "kitchen_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}